split code into multiple files and refactored from cjs to mjs
